requirements.txt
transformers
sentence-transformers
faiss-cpu
pinecone-client
PyPDF2
